loader: Split logger recorders into separate file.
Reduces API surface area
